adding example for MD README that we can use as template
i believe it is kind of important all the README to see the same, i also
with this humble template, i explain how i will improve the README, (i
will let the description area for you, (for Bucky Roberts) to explain
the course and i will make a nice documentation for the syntax that you
use in your lessons). Do not worry, i changed only empty README.md and
only in the Arrays folder.
If you agree tell me to start to fill this babies with Knowledge :)